chore: db mybatis 테스트 위한 컨트롤러, 서비스 추가 및 필요한 파일 수정

+++ src/main/java/fakehunters/backend/audio/mapper/AudioAnalysisResultMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Param;
 @Mapper
 public interface AudioAnalysisResultMapper {
 
-    int insert(AudioAnalysisResultInsertParam param);
+    long insert(AudioAnalysisResultInsertParam param);
 
     AudioAnalysisResult findById(@Param("id") long id);
 }
